requirements.txt
numpy==1.17.0
scipy==1.6.2
scikit-learn==1.0.2
keras==2.4.3
tensorflow-gpu==2.4.1